+++ src/cppNGS/StatisticsReads.h
@@ -33,12 +33,12 @@ public:
 
 private:
 	QMutex mutex_;
-	int c_forward_;
-	int c_reverse_;
+	long long c_forward_;
+	long long c_reverse_;
 	QSet<int> read_lengths_;
     long long bases_sequenced_;
-	double c_read_q20_;
-	double c_base_q30_;
+	long long c_read_q20_;
+	long long c_base_q30_;
 	QVector<Pileup> pileups_;
 	QVector<double> qualities1_;
 	QVector<double> qualities2_;
